Refactor: 입력 문자를 피연산자로 만드는 Operands 클래스와 숫자를 더하는 Calculator 클래스로 역할 분리
